requirements.txt
groq>=0.4.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
numpy>=1.24.0
